+++ pkg/tasks/prometheus_user_workload.go
@@ -139,6 +139,16 @@ func (t *PrometheusUserWorkloadTask) create() error {
 		return errors.Wrap(err, "reconciling UserWorkload Prometheus Service failed")
 	}
 
+	svc, err = t.factory.PrometheusUserWorkloadServiceThanosSidecar()
+	if err != nil {
+		return errors.Wrap(err, "initializing UserWorkload Thanos sidecar Service failed")
+	}
+
+	err = t.client.CreateOrUpdateService(svc)
+	if err != nil {
+		return errors.Wrap(err, "reconciling UserWorkload Thanos sidecar Service failed")
+	}
+
 	grpcTLS, err := t.factory.GRPCSecret()
 	if err != nil {
 		return errors.Wrap(err, "initializing UserWorkload Prometheus GRPC secret failed")
@@ -201,10 +211,34 @@ func (t *PrometheusUserWorkloadTask) create() error {
 	}
 
 	err = t.client.CreateOrUpdateServiceMonitor(smp)
-	return errors.Wrap(err, "reconciling UserWorkload Prometheus ServiceMonitor failed")
+	if err != nil {
+		return errors.Wrap(err, "reconciling UserWorkload Prometheus ServiceMonitor failed")
+	}
+
+	smt, err := t.factory.PrometheusUserWorkloadThanosSidecarServiceMonitor()
+	if err != nil {
+		return errors.Wrap(err, "initializing UserWorkload Thanos sidecar ServiceMonitor failed")
+	}
+
+	err = t.client.CreateOrUpdateServiceMonitor(smt)
+	if err != nil {
+		return errors.Wrap(err, "reconciling UserWorkload Thanos sidecar ServiceMonitor failed")
+	}
+
+	return nil
 }
 
 func (t *PrometheusUserWorkloadTask) destroy() error {
+	smt, err := t.factory.PrometheusUserWorkloadThanosSidecarServiceMonitor()
+	if err != nil {
+		return errors.Wrap(err, "initializing UserWorkload Thanos sidecar ServiceMonitor failed")
+	}
+
+	err = t.client.DeleteServiceMonitor(smt)
+	if err != nil {
+		return errors.Wrap(err, "deleting UserWorkload Thanos sidecar ServiceMonitor failed")
+	}
+
 	smp, err := t.factory.PrometheusUserWorkloadPrometheusServiceMonitor()
 	if err != nil {
 		return errors.Wrap(err, "initializing UserWorkload Prometheus ServiceMonitor failed")
@@ -254,6 +288,16 @@ func (t *PrometheusUserWorkloadTask) destroy() error {
 	svc, err := t.factory.PrometheusUserWorkloadService()
 	if err != nil {
 		return errors.Wrap(err, "initializing UserWorkload Prometheus Service failed")
+	}
+
+	err = t.client.DeleteService(svc)
+	if err != nil {
+		return errors.Wrap(err, "deleting UserWorkload Prometheus Service failed")
+	}
+
+	svc, err = t.factory.PrometheusUserWorkloadServiceThanosSidecar()
+	if err != nil {
+		return errors.Wrap(err, "initializing UserWorkload Thanos sidecar Service failed")
 	}
 
 	err = t.client.DeleteService(svc)